+++ config/branch/config.go
@@ -2,7 +2,7 @@ package branch
 
 import (
 	"github.com/gitamix/lint/config/branch/name"
-	"github.com/gitamix/lint/config/ticket"
+	"github.com/gitamix/lint/config/task"
 )
 
 // Config represents configuration for commit branch behavior.
@@ -10,8 +10,8 @@ type Config struct {
 	// name stores the configuration for branch naming rules.
 	name name.Config
 
-	// tkt stores the configuration for ticket integration.
-	tkt ticket.Config
+	// tkt stores the configuration for task integration.
+	tkt task.Config
 }
 
 // NewConfig creates a new Config instance
@@ -29,7 +29,7 @@ func (c Config) Name() name.Config {
 	return c.name
 }
 
-// Ticket returns the configuration for ticket integration.
-func (c Config) Ticket() ticket.Config {
+// Task returns the configuration for task integration.
+func (c Config) Task() task.Config {
 	return c.tkt
 }