+++ config/ticket/doc.go
@@ -1,3 +1,0 @@
-// Package ticket provides types and functions
-// for working with ticket (issue) configuration.
-package ticket