+++ config/ticket/id/doc.go
@@ -1,3 +1,0 @@
-// Package id provides types and functions
-// for working with ticket (issue) identifier configuration.
-package id